+++ lib/exporto_api/objects/label_method_response.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ExportoAPI
+  module Objects
+    class LabelMethodResponse < Base; end
+  end
+end