+++ lib/exporto_api/resources/label_method_resource.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module ExportoAPI
+  class LabelMethodResource < Resource
+    def all
+      Objects::LabelMethodResponse.new(get_request("label/methods"))
+    end
+  end
+end